Add ability to specify F_link independently
Adds F_link_method to bolo_config which allows user to select whether
to compute F_link from beta, Tbath, and Tc, or specify its value as
part of the channel configuration. Backwards compatibility is
maintained by defaulting to the usual computation of F_link when
F_link_method is not specified.