Add band file option for absorption and reflection
Adds the option to specify band files for absorption ('Bespoke' optical
elements) and reflection (all optical elements). These properties are
assumed to be defined by files when they are specified as string types
(i.e. filenames), thus maintaining compatibility with existing config
files in which they are specified as numerical or list types.